messaging-service/pyproject.toml
[tool.poetry]
name = "arxiv-messaging"
version = "0.1.0"
description = "arXiv messaging service"
authors = ["arXiv dev <developers@arxiv.org>"]
license = "BSD-3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
ruamel-yaml = "^0.18.6"
python-json-logger = "^2.0.7"
python-dateutil = "^2.8.2"
httpx = "^0.28.1"
cachetools = "^5.5.0"
asgi-correlation-id = "^4.3.4"
asgi-logger = "^0.1.0"
python-multipart = "^0.0.20"
google-cloud-pubsub = "^2.31.1"
schedule = "^1.2.2"
firebase-admin = "^7.1.0"
structlog = "^24.1.0"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
pytest = "^8.3.5"
pytest-mock = "^3.12.0"
pytest-cov = "^4.1.0"
types-pymysql = "^1.1.0.20250516"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
